+++ app/controllers/api/teams.php
@@ -243,7 +243,7 @@ App::delete('/v1/teams/:teamId')
         }
 
         $events
-            ->setParam('payload', $response->output($team, Response::MODEL_TEAM))
+            ->setParam('eventData', $response->output($team, Response::MODEL_TEAM))
         ;
 
         $response->noContent();
@@ -344,6 +344,7 @@ App::post('/v1/teams/:teamId/memberships')
                     'registration' => \time(),
                     'reset' => false,
                     'name' => $name,
+                    'sessions' => [],
                     'tokens' => [],
                 ], ['email' => $email]);
             } catch (Duplicate $th) {
@@ -612,10 +613,11 @@ App::patch('/v1/teams/:teamId/memberships/:inviteId/status')
         $expiry = \time() + Auth::TOKEN_EXPIRATION_LOGIN_LONG;
         $secret = Auth::tokenGenerator();
         $session = new Document(array_merge([
-            '$collection' => Database::SYSTEM_COLLECTION_TOKENS,
+            '$collection' => Database::SYSTEM_COLLECTION_SESSIONS,
             '$permissions' => ['read' => ['user:'.$user->getId()], 'write' => ['user:'.$user->getId()]],
             'userId' => $user->getId(),
-            'type' => Auth::TOKEN_TYPE_LOGIN,
+            'provider' => Auth::SESSION_PROVIDER_EMAIL,
+            'providerUid' => $user->getAttribute('email'),
             'secret' => Auth::hash($secret), // One way hash encryption to protect DB leak
             'expire' => $expiry,
             'userAgent' => $request->getUserAgent('UNKNOWN'),
@@ -623,7 +625,7 @@ App::patch('/v1/teams/:teamId/memberships/:inviteId/status')
             'countryCode' => ($record) ? \strtolower($record['country']['iso_code']) : '--',
         ], $detector->getOS(), $detector->getClient(), $detector->getDevice()));
 
-        $user->setAttribute('tokens', $session, Document::SET_TYPE_APPEND);
+        $user->setAttribute('sessions', $session, Document::SET_TYPE_APPEND);
 
         Authorization::setRole('user:'.$userId);
 
@@ -728,7 +730,7 @@ App::delete('/v1/teams/:teamId/memberships/:inviteId')
         ;
 
         $events
-            ->setParam('payload', $response->output($membership, Response::MODEL_MEMBERSHIP))
+            ->setParam('eventData', $response->output($membership, Response::MODEL_MEMBERSHIP))
         ;
 
         $response->noContent();