+++ src/Appwrite/Auth/Auth.php
@@ -28,10 +28,16 @@ class Auth
     /**
      * Token Types.
      */
-    const TOKEN_TYPE_LOGIN = 1;
+    const TOKEN_TYPE_LOGIN = 1; // Deprecated
     const TOKEN_TYPE_VERIFICATION = 2;
     const TOKEN_TYPE_RECOVERY = 3;
     const TOKEN_TYPE_INVITE = 4;
+
+    /**
+     * Session Providers.
+     */
+    const SESSION_PROVIDER_EMAIL = 'email';
+    const SESSION_PROVIDER_ANONYMOUS = 'anonymous';
 
     /**
      * Token Expiration times.
@@ -201,6 +207,29 @@ class Auth
                 $token->getAttribute('secret') === self::hash($secret) &&
                 $token->getAttribute('expire') >= \time()) {
                 return (string)$token->getId();
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Verify session and check that its not expired.
+     *
+     * @param array  $sessions
+     * @param string $secret
+     *
+     * @return bool|string
+     */
+    public static function sessionVerify(array $sessions, string $secret)
+    {
+        foreach ($sessions as $session) { /** @var Document $session */
+            if ($session->isSet('secret') &&
+                $session->isSet('expire') &&
+                $session->isSet('provider') &&
+                $session->getAttribute('secret') === self::hash($secret) &&
+                $session->getAttribute('expire') >= \time()) {
+                return (string)$session->getId();
             }
         }